# openapi.yml
openapi: 3.0.3
info:
  title: Hyper Fleet Native API
  description: |
    Hyper Fleet Native API provides comprehensive management of virtual machines (sandboxes), 
    file system operations, execution environments, and organizational resources.

    All endpoints except `/api/register` require the `X-API-Key` header for authentication.
  version: 1.0.0
  contact:
    name: Hyper Fleet Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:33944/api
    description: Local development server
  - url: https://api.hyperfleet.io/api
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key for authentication (format: hf_...)"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        details:
          type: string
          example: Additional error details

    # Registration
    RegisterRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: Admin User
        email:
          type: string
          format: email
          maxLength: 254
          example: admin@example.com
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: User registered successfully
        data:
          type: object
          properties:
            orgId:
              type: string
              example: 65ae1234567890abcdef1234
            apiKey:
              type: string
              example: hf_1234567890abcdef

    # Organization
...
    ExecRequest:
      type: object
      required:
        - command
      properties:
        command:
          type: string
          example: ls -la /root
        args:
          type: array
          items:
            type: string
          example: [-la, /root]
        timeout:
          type: integer
          default: 30
          example: 30
          description: Timeout in seconds (ignored if background=true)
        env:
          type: object
          additionalProperties:
            type: string
          example:
            PATH: /usr/bin:/bin
        cwd:
          type: string
          example: /root
        background:
          type: boolean
          default: false
          description: If true, starts process in background and returns PID immediately
...
paths:
  /register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: |
        Create a new user account and default organization.
        This is the only public endpoint - no authentication required.
        **Important:** Save the returned API key immediately as it's only shown once.
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /sandboxes/{id}/exec-stream:
    post:
      tags:
        - Execution
      summary: Execute command (SSE stream)
      description: |
        Execute a command and stream output as Server-Sent Events.
        Events: `stdout`, `stderr`, and final `exit`.
      operationId: execCommandStream
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecRequest"
      responses:
        "200":
          description: SSE stream of command output
          content:
            text/event-stream:
              schema:
                type: string
              example: |-
                event: stdout
                data: Hello from exec

                event: stderr
                data: warning: something happened

                event: exit
                data: {"exitCode":0}
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...